Added OnClickDisappear Class
Depending on its bool state, the class will change its behaviour:
Is EnableMode on? Disable all objects spriterenderer and if you click on it enable it
Is EnableMode off? Enable Objects and SpriteRenderer, if clicked, disable object.
Is SwitchMode on? Disable all objects in the array except the first one. Is the first one clicked then disable this one and enable the next one.