Added Indicator for most minigames
In the prefabs folder is now a shrunken down butterfly with an animation called Indicator. Add this as a child to any minigame and everything should work smoothly. For the make objects disappear => Always add indicator as the first child of a clickable object. This way the indicator will disappear once its clicked